Add __init__.py files to tests
Previously, the tests were simple '.py' files. This meant that the tests
were not using fully qualified names and having duplicate tests might
cause a conflict like:

import file mismatch: imported module 'test_core' has this __file__
attribute [...] which is not the same as the test file we want to
collect [...]. HINT: remove __pycache__ / .pyc files and/or use a unique
basename for your test file modules

To avoid this problem, we make the tests/ folder a proper package.